Added win condition and win screen. Added "Play Again?" option to the winScreen.
