+++ core/src/test/java/edu/bsu/storygame/core/model/GameContextTest.java
@@ -1,7 +1,11 @@
 package edu.bsu.storygame.core.model;
 
+import com.google.common.collect.Lists;
 import edu.bsu.storygame.core.MonsterGame;
 import org.junit.Test;
+import tripleplay.util.Colors;
+
+import java.util.List;
 
 import static org.junit.Assert.assertNull;
 import static org.mockito.Mockito.mock;
@@ -10,9 +14,12 @@ public final class GameContextTest {
 
     @Test
     public void testEncounter_isClearedAtEndOfRound() {
+        final List<Skill> skills = Lists.newArrayList();
         GameContext context = new GameContext(mock(MonsterGame.class),
-                mock(Player.class),
-                mock(Player.class));
+                new Player.Builder().name("Jack")
+                        .color(Colors.CYAN).skills(skills).build(),
+                new Player.Builder().name("Also Jack")
+                        .color(Colors.CYAN).skills(skills).build());
         context.encounter.update(mock(Encounter.class));
         context.phase.update(Phase.END_OF_ROUND);
         assertNull(context.encounter.get());
